move DSMC filter functions to different file
